rename pkg2
