+++ internal/components/console_panel.go
@@ -1,8 +1,8 @@
 package components
 
 import (
-	"github.com/sgrumley/gotex/pkgv2/ansi"
-	"github.com/sgrumley/gotex/pkgv2/runner"
+	"github.com/sgrumley/gotex/pkg/ansi"
+	"github.com/sgrumley/gotex/pkg/runner"
 
 	"github.com/gdamore/tcell/v2"
 	"github.com/rivo/tview"
