+++ pkg/config/config_test.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/sgrumley/gotex/pkgv2/config"
+	"github.com/sgrumley/gotex/pkg/config"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
