+++ pkg/models/file.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/sgrumley/gotex/pkgv2/runner"
+	"github.com/sgrumley/gotex/pkg/runner"
 )
 
 var _ Node = (*File)(nil)
