+++ pkg/models/package.go
@@ -3,7 +3,7 @@ package models
 import (
 	"strings"
 
-	"github.com/sgrumley/gotex/pkgv2/runner"
+	"github.com/sgrumley/gotex/pkg/runner"
 )
 
 var _ Node = (*Package)(nil)
