+++ pkg/scanner/ast.go
