+++ pkg/scanner/tree.go
